Move `Footer` to ~/components
