+++ src/pages/SponsorshipsPage.tsx
@@ -17,7 +17,6 @@ import Tabs, { Tab } from '~/shared/components/Tabs'
 import Button from '~/shared/components/Button'
 import { ScrollTableCore } from '~/shared/components/ScrollTable/ScrollTable'
 import { useWalletAccount } from '~/shared/stores/wallet'
-import Footer from '~/shared/components/Layout/Footer'
 import CreateSponsorshipModal from '~/modals/CreateSponsorshipModal'
 import { LoadMoreButton } from '~/components/LoadMore'
 import useIsMounted from '~/shared/hooks/useIsMounted'
