+++ src/shared/components/Layout/index.tsx
@@ -6,8 +6,8 @@ import cx from 'classnames'
 import { COLORS, MAX_BODY_WIDTH } from '~/shared/utils/styled'
 import useScrollToTop from '~/shared/hooks/useScrollToTop'
 import useCurrentLocation from '~/shared/hooks/useCurrentLocation'
+import Footer from '~/components/Footer'
 import Nav from './Nav'
-import Footer from './Footer'
 import styles from './layout.pcss'
 
 const Inner = styled.div<{ $gray?: boolean }>`
